[#207] Clean up serializer tests, schema helper
